+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Fixed
+- Fixing struct issue in 14.ADCSensorExample
 
 ## [1.0.0] - 2025-04-15